+++ spec/integration/rest/studio/v1/flow/engagement.spec.js
@@ -34,8 +34,8 @@ describe('Engagement', function() {
     function() {
       holodeck.mock(new Response(500, '{}'));
 
-      var promise = client.preview.studio.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
-                                         .engagements.list();
+      var promise = client.studio.v1.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+                                    .engagements.list();
       promise = promise.then(function() {
         throw new Error('failed');
       }, function(error) {
@@ -44,7 +44,7 @@ describe('Engagement', function() {
       promise.done();
 
       var solution = {flowSid: 'FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'};
-      var url = _.template('https://preview.twilio.com/Studio/Flows/<%= flowSid %>/Engagements')(solution);
+      var url = _.template('https://studio.twilio.com/v1/Flows/<%= flowSid %>/Engagements')(solution);
 
       holodeck.assertHasRequest(new Request({
         method: 'GET',
@@ -58,9 +58,9 @@ describe('Engagement', function() {
           'meta': {
               'previous_page_url': null,
               'next_page_url': null,
-              'url': 'https://preview.twilio.com/Studio/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements?PageSize=50&Page=0',
+              'url': 'https://studio.twilio.com/v1/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements?PageSize=50&Page=0',
               'page': 0,
-              'first_page_url': 'https://preview.twilio.com/Studio/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements?PageSize=50&Page=0',
+              'first_page_url': 'https://studio.twilio.com/v1/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements?PageSize=50&Page=0',
               'page_size': 50,
               'key': 'engagements'
           },
@@ -69,8 +69,8 @@ describe('Engagement', function() {
 
       holodeck.mock(new Response(200, body));
 
-      var promise = client.preview.studio.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
-                                         .engagements.list();
+      var promise = client.studio.v1.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+                                    .engagements.list();
       promise = promise.then(function(response) {
         expect(response).toBeDefined();
       }, function() {
@@ -84,8 +84,8 @@ describe('Engagement', function() {
     function() {
       holodeck.mock(new Response(500, '{}'));
 
-      var promise = client.preview.studio.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
-                                         .engagements('FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').fetch();
+      var promise = client.studio.v1.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+                                    .engagements('FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').fetch();
       promise = promise.then(function() {
         throw new Error('failed');
       }, function(error) {
@@ -97,7 +97,7 @@ describe('Engagement', function() {
         flowSid: 'FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
         sid: 'FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };
-      var url = _.template('https://preview.twilio.com/Studio/Flows/<%= flowSid %>/Engagements/<%= sid %>')(solution);
+      var url = _.template('https://studio.twilio.com/v1/Flows/<%= flowSid %>/Engagements/<%= sid %>')(solution);
 
       holodeck.assertHasRequest(new Request({
         method: 'GET',
@@ -117,16 +117,16 @@ describe('Engagement', function() {
           'context': {},
           'date_created': '2017-11-06T12:00:00Z',
           'date_updated': null,
-          'url': 'https://preview.twilio.com/Studio/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'url': 'https://studio.twilio.com/v1/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
-              'steps': 'https://preview.twilio.com/Studio/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Steps'
+              'steps': 'https://studio.twilio.com/v1/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Steps'
           }
       });
 
       holodeck.mock(new Response(200, body));
 
-      var promise = client.preview.studio.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
-                                         .engagements('FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').fetch();
+      var promise = client.studio.v1.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+                                    .engagements('FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').fetch();
       promise = promise.then(function(response) {
         expect(response).toBeDefined();
       }, function() {
@@ -141,8 +141,8 @@ describe('Engagement', function() {
       holodeck.mock(new Response(500, '{}'));
 
       var opts = {to: '+15558675310', from: '+15017122661'};
-      var promise = client.preview.studio.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
-                                         .engagements.create(opts);
+      var promise = client.studio.v1.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+                                    .engagements.create(opts);
       promise = promise.then(function() {
         throw new Error('failed');
       }, function(error) {
@@ -151,7 +151,7 @@ describe('Engagement', function() {
       promise.done();
 
       var solution = {flowSid: 'FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'};
-      var url = _.template('https://preview.twilio.com/Studio/Flows/<%= flowSid %>/Engagements')(solution);
+      var url = _.template('https://studio.twilio.com/v1/Flows/<%= flowSid %>/Engagements')(solution);
 
       var values = {To: '+15558675310', From: '+15017122661', };
       holodeck.assertHasRequest(new Request({
@@ -164,7 +164,7 @@ describe('Engagement', function() {
   it('should generate valid create response',
     function() {
       var body = JSON.stringify({
-          'url': 'https://preview.twilio.com/Studio/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'url': 'https://studio.twilio.com/v1/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'sid': 'FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'flow_sid': 'FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
@@ -179,15 +179,15 @@ describe('Engagement', function() {
           'date_created': '2015-07-30T20:00:00Z',
           'date_updated': '2015-07-30T20:00:00Z',
           'links': {
-              'steps': 'https://preview.twilio.com/Studio/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Steps'
+              'steps': 'https://studio.twilio.com/v1/Flows/FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Engagements/FNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Steps'
           }
       });
 
       holodeck.mock(new Response(201, body));
 
       var opts = {to: '+15558675310', from: '+15017122661'};
-      var promise = client.preview.studio.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
-                                         .engagements.create(opts);
+      var promise = client.studio.v1.flows('FWaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+                                    .engagements.create(opts);
       promise = promise.then(function(response) {
         expect(response).toBeDefined();
       }, function() {