+++ spec/integration/rest/video/v1/recording.spec.js
@@ -65,6 +65,7 @@ describe('Recording', function() {
           'duration': 0,
           'container_format': 'mka',
           'codec': 'OPUS',
+          'track_name': 'A name',
           'grouping_sids': {
               'room_sid': 'RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
           },
@@ -149,6 +150,7 @@ describe('Recording', function() {
                   'duration': 10,
                   'container_format': 'mka',
                   'codec': 'OPUS',
+                  'track_name': 'A name',
                   'grouping_sids': {
                       'room_sid': 'RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                       'participant_sid': 'PAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'